% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dS.R
\name{calc_dS}
\alias{calc_dS}
\title{Calculation of delta network smoothing indexes}
\usage{
calc_dS(X0, A, classes, eps = 1, ...)
}
\arguments{
\item{X0}{input matrix}

\item{A}{adjancency matrix}

\item{classes}{numeric vector of \{1,2\}}

\item{eps}{numeric value}

\item{...}{additional parameteres of network_propagation}
}
\value{
\code{data.frame} with delta gene network smoothing indexes
}
\description{
Calculation of delta network smoothing indexes
}
