% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Sp.R
\name{calc_Sp}
\alias{calc_Sp}
\title{Calculation of permutation-adjusted network smoothing index}
\usage{
calc_Sp(X0, A, eps = 1, k = 99, mc.cores = 1, ...)
}
\arguments{
\item{X0}{input matrix}

\item{A}{adjancency matrix}

\item{eps}{numeric value}

\item{k}{number of permutations}

\item{mc.cores}{number of cores}

\item{...}{additional parameteres of network_propagation}
}
\value{
\code{data.frame} with X0, Xs, S, p and Sp
}
\description{
Calculation of permutation-adjusted network smoothing index
}
