% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dS.R
\name{dS}
\alias{dS}
\title{Delta Network Smoothing Index}
\usage{
dS(X0, Xs, classes, eps = 1)
}
\arguments{
\item{X0}{input matrix}

\item{Xs}{smoothed matrix}

\item{classes}{vector of classes, values must be equal to 1 or 2}

\item{eps}{numeric value}
}
\value{
named vector of delta smoothing indexes
}
\description{
Delta Network Smoothing Index
}
