% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_map.R
\name{linear_map}
\alias{linear_map}
\title{Linear map}
\usage{
linear_map(x, y.min, y.max)
}
\arguments{
\item{x}{numeric vector}

\item{y.min}{min value of the new distribution}

\item{y.max}{max value of the new distribution}
}
\description{
Linear map
}
