% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_propagation.R
\name{network_propagation}
\alias{network_propagation}
\title{Network propagation}
\usage{
network_propagation(
  F0,
  A,
  alpha = 0.7,
  nMax = 10000,
  eps = 1e-06,
  final.smooth = FALSE,
  all.steps = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{F0}{vector or matrix composed of column vectors with initial distribution of information}

\item{A}{normalized adjacency matrix}

\item{alpha}{numeric, the smothing factor}

\item{nMax}{numeric, maximum number of iterations}

\item{eps}{numeric, the iteration will stop when the maximum difference between matrix Ft between two consecutive iteraction is smaller than \code{eps}}

\item{final.smooth}{TRUE/FALSE, whether to do the final step of smoothing}

\item{all.steps, }{TRUE/FALSE, whether to store all steps}

\item{verbose, }{TRUE/FALSE}
}
\value{
a list with:
\itemize{
\item{\code{Ft}}{ the smoothed matrix;}
\item{\code{eps}}{ see above;}
\item{\code{max.abs.diff}}{ max(abs(F_t) - abs(F_{t-1}));}
\item{\code{Ft.all}}{ transient Ft matrices.}
}
}
\description{
Network propagation
}
