% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_adj_mat.R
\name{normalize_adj_mat}
\alias{normalize_adj_mat}
\title{Symmetric Normalization of Adjancency Matrix}
\usage{
normalize_adj_mat(A)
}
\arguments{
\item{A}{adjacency matrix}
}
\value{
normalized adjacency matrix
}
\description{
Symmetric Normalization of Adjancency Matrix
}
\details{
This function applies the following normalization: a_ij' = a_ij / sqrt(d_i d_j), where d_i is the degree of vertex i
}
