% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NR.R
\name{NR}
\alias{NR}
\title{Local Monte Carlo p values}
\usage{
NR(edge_list, ranked_vector, k = 99, mc.cores = 1)
}
\arguments{
\item{edge_list}{edge list}

\item{ranked_vector}{scores vector; it must have the same names and size of the vertices of G}

\item{k}{number of permutations}

\item{mc.cores}{number of cores}
}
\description{
Local Monte Carlo p values
}
