% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsi.R
\name{nsi}
\alias{nsi}
\title{Network Smoothing Index}
\usage{
nsi(x, y, eps = 1)
}
\arguments{
\item{x}{input matrix/vector}

\item{y}{smoothed matrix/vector}

\item{eps}{numeric value}
}
\value{
data.frame with X0, Xs, and S
}
\description{
Network Smoothing Index
}
