% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega.R
\name{omega}
\alias{omega}
\title{Calculation of omega function}
\usage{
omega(edge_list, dS)
}
\arguments{
\item{edge_list}{edge list}

\item{dS}{scores vector; it must have the same names and size of the vertices of G}
}
\description{
Calculation of omega function
}
