% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_perm.R
\name{omega_perm}
\alias{omega_perm}
\title{Calculation of permuted omega function}
\usage{
omega_perm(idx, edge_list, dS)
}
\arguments{
\item{idx}{vector of randomly resampled dS scores; the names must be the same as dS's}

\item{edge_list}{edge list}

\item{dS}{scores vector; it must have the same names and size of the vertices of G}
}
\description{
Calculation of permuted omega function
}
